PRIMARY KEYwo

+++ springboot/src/test/resources/schema-product.sql
@@ -7,7 +7,7 @@ DROP TABLE IF EXISTS asakatu_event;
 DROP TABLE IF EXISTS user;
 
 CREATE TABLE IF NOT EXISTS user(
-    user_id INT PRIMARY KEY AUTO_INCREMENT,
+    user_id BIGINT(20) PRIMARY KEY AUTO_INCREMENT,
     username VARCHAR(50) NOT NULL,
     email VARCHAR(255) NOT NULL,
     password VARCHAR(50) NOT NULL,
@@ -20,7 +20,7 @@ CREATE TABLE IF NOT EXISTS user(
 
 /* TODO asakatu_eventとasakatu_event_idの名称を変更する jojo */
 CREATE TABLE IF NOT EXISTS asakatu_event (
-     asakatu_event_id INT PRIMARY KEY AUTO_INCREMENT,
+     asakatu_event_id BIGINT(20) PRIMARY KEY AUTO_INCREMENT,
      start_date DATETIME,
      duration DOUBLE,
      address VARCHAR(255),
@@ -31,26 +31,26 @@ CREATE TABLE IF NOT EXISTS asakatu_event (
 );
 
 CREATE TABLE IF NOT EXISTS user_event_association (
-    user_id INT,
-    asakatu_event_id INT,
+    user_id BIGINT(20),
+    asakatu_event_id BIGINT(20),
     event_canceled BOOLEAN,
     FOREIGN KEY (user_id) REFERENCES user(user_id),
     FOREIGN KEY (asakatu_event_id) REFERENCES asakatu_event(asakatu_event_id)
 );
 
 CREATE TABLE IF NOT EXISTS user_status_master (
-    user_status_master_id INT PRIMARY KEY AUTO_INCREMENT,
+    user_status_master_id BIGINT(20) PRIMARY KEY AUTO_INCREMENT,
     user_status_content VARCHAR(50),
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
 );
 
 CREATE TABLE IF NOT EXISTS user_status (
-    user_status_id INT PRIMARY KEY AUTO_INCREMENT,
-    user_status_master_id INT,
+    user_status_id BIGINT(20) PRIMARY KEY AUTO_INCREMENT,
+    user_status_master_id BIGINT(20),
     user_status_comment VARCHAR(255),
-    user_id INT,
-    asakatu_event_id INT,
+    user_id BIGINT(20),
+    asakatu_event_id BIGINT(20),
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     FOREIGN KEY (user_id) REFERENCES user(user_id),
@@ -59,8 +59,8 @@ CREATE TABLE IF NOT EXISTS user_status (
 );
 
 CREATE TABLE IF NOT EXISTS user_group (
-    user_group_id INT PRIMARY KEY AUTO_INCREMENT,
-    order_user_id INT,
+    user_group_id BIGINT(20) PRIMARY KEY AUTO_INCREMENT,
+    order_user_id BIGINT(20),
     product_key VARCHAR(255),
     description VARCHAR(255),
     created_at DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
@@ -69,8 +69,8 @@ CREATE TABLE IF NOT EXISTS user_group (
 );
 
 CREATE TABLE IF NOT EXISTS user_group_association (
-    user_id INT,
-    user_group_id INT,
+    user_id BIGINT(20),
+    user_group_id BIGINT(20),
     FOREIGN KEY (user_id) REFERENCES user(user_id),
     FOREIGN KEY (user_group_id) REFERENCES user_group(user_group_id)
 );
